% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnote.symbols.R
\docType{data}
\name{footnote.symbols}
\alias{footnote.symbols}
\title{Foot note symbols}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{symbol}{Footnote symbol}
\item{name}{Name of footnote symbol}
}
}
\usage{
footnote.symbols
}
\description{
A dataset containing footnote symbols.
}
\keyword{datasets}
