% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountry.R
\name{getCountry}
\alias{getCountry}
\title{Get country name from ISO 3-character code}
\usage{
getCountry(iso3, shortest = FALSE)
}
\arguments{
\item{iso3}{A vector of ISO 3-character codes}

\item{shortest}{A logical indicating whether the shortest possible name should be returned (i.e., US = United States)}
}
\value{
A vector of country names
}
\description{
Get country name from ISO 3-character code
}
\examples{
iso3 <- c("CHN", "SOM", "USA")
getCountry(iso3)
getCountry(iso3, shortest = TRUE)
}
