% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operands.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Value exclusion}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{A vector containing the values to check for exclusion}

\item{y}{A vector containing the values to be checked against}
}
\value{
A logical indicating if x was not found in y
}
\description{
\verb{\%notin\%} is the opposite of the \code{\%in\%} operator. It returns a logical vector indicating if the left operand is not a match in the right operand.
}
\examples{
x <- seq(1, 10, 1)
y <- seq(1, 10, 2)
x \%notin\% y
}
