% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupColor.R
\name{groupColor}
\alias{groupColor}
\title{Title}
\usage{
groupColor(
  n,
  names = NULL,
  palettes = NULL,
  color.and.fill = FALSE,
  show = FALSE
)
}
\arguments{
\item{n}{Number of shades needed by group}

\item{names}{Names of shades}

\item{palettes}{ggsci material design color palettes}

\item{color.and.fill}{Logical indicator of whether to both fill colors and darkened colors (for borders)}

\item{show}{Logical indicators of whether to show output colors}
}
\value{
Vector of colors
}
\description{
Title
}
