% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupColorInfo.R
\name{groupColorInfo}
\alias{groupColorInfo}
\title{Color shades needed by group}
\usage{
groupColorInfo(df, group, subgroup)
}
\arguments{
\item{df}{Data frame}

\item{group}{Grouping variable for color family (e.g., reds)}

\item{subgroup}{Subgroup variable for shades (e.g., burgundy, cherry, scarlett)}
}
\value{
List of two with the number of unique color shades needed by group with subgroup labels
}
\description{
For use with groupColorFunct
}
\examples{
cars <- cars \%>\% dplyr::mutate(across(everything(), ~plyr::round_any(., f = floor, accuracy = 10)))
groupColorInfo(cars, "speed", "dist")
}
