% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelAuto.R
\name{labelAuto}
\alias{labelAuto}
\title{Pretty scientific scale for visualization}
\usage{
labelAuto(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A character vector for formatted numbers
}
\description{
Pretty scientific scale for visualization
}
\examples{
ex1 <- sort(rnorm(n = 10, mean = 0, sd = 1))
labelAuto(ex1)
ex2 <- sort(rnorm(n = 10, mean = 0, sd = 1e3))
labelAuto(ex2)
ex3 <- sort(rnorm(n = 10, mean = 0, sd = 1e6))
labelAuto(ex3)
ex4 <- sort(rnorm(n = 10, mean = 0, sd = 1e8))
labelAuto(ex4)
}
