% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCaption.R
\name{makeCaption}
\alias{makeCaption}
\title{Create single caption from a list of captions}
\usage{
makeCaption(x)
}
\arguments{
\item{x}{A list of captions}
}
\value{
A compiled caption string
}
\description{
Create single caption from a list of captions
}
\examples{
captions <- list("A" = "* This is caption A",
"B" = "Note: This is caption B")
makeCaption(captions)
}
