% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMoney.R
\name{makeMoney}
\alias{makeMoney}
\title{Format currency values}
\usage{
makeMoney(x, currency = "$", cents = FALSE)
}
\arguments{
\item{x}{A numeric vector representing currency}

\item{currency}{The currency character(s) to include (use "" if no currency character should be included)}

\item{cents}{A logical indicating if the values should include cents}
}
\value{
A character vector of formatted numerics
}
\description{
This function formats currency values. It is helpful when producing values that will be included as labels in graphs or values in display tables.
}
\examples{
x <- runif(10, min = 0, max = 1000)
makeMoney(x, currency = "$", cents = TRUE)
makeMoney(x, currency = "USD ", cents = FALSE)
makeMoney(x, currency = "", cents = FALSE)
}
