% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePath.R
\name{makePath}
\alias{makePath}
\title{Make file path from folder and filename}
\usage{
makePath(folder, filename, extension = NULL)
}
\arguments{
\item{folder}{A string of text denoting the folder to save the file to}

\item{filename}{A string of text denoting the filename}

\item{extension}{An option string of text denoting the file extension if not included in the filename}
}
\value{
A file path
}
\description{
Make file path from folder and filename
}
\examples{
makePath(folder = "output-data", filename = "file.Rda")
makePath(folder = "output-data", filename = "file", extension = ".Rda")
}
