% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRound.R
\name{makeRound}
\alias{makeRound}
\title{Format and round numbers}
\usage{
makeRound(x, accuracy = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{accuracy}{The number to round to (e.g., 0.05, 5)}
}
\value{
A character vector of formatted numbers
}
\description{
This function rounds and formats numbers. It is helpful when producing values that will be included as labels in graphs or values in display tables.
}
\examples{
x <- runif(10, min = 0, max = 1000)
makeRound(x, accuracy = 0.01)
makeRound(x, accuracy = 10)
}
