% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Print a notification in the console}
\usage{
notify(x)
}
\arguments{
\item{x}{Text to notify}
}
\value{
Printed notification text
}
\description{
Print a notification in the console
}
\examples{
notify("Loop complete")
}
