% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarahLoad.R
\name{sarahLoad}
\alias{sarahLoad}
\title{Loads .rda objects from the project's output data folder}
\usage{
sarahLoad(object, folder = "03-output_data")
}
\arguments{
\item{object}{Name of object to load}

\item{folder}{Output data folder}
}
\value{
Loads object into environment
}
\description{
Loads .rda objects from the project's output data folder
}
