% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarahSave.R
\name{sarahSave}
\alias{sarahSave}
\title{Saves objects to the project's output data folder}
\usage{
sarahSave(object, folder = "3 Output data", type = ".rda")
}
\arguments{
\item{object}{Object to save}

\item{folder}{Folder to save the object to}

\item{type}{Character string of file type; ".rda" or ".csv}
}
\value{
Notification of saved object
}
\description{
Saves objects to the project's output data folder
}
