% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveGGplot.R
\name{saveGGplot}
\alias{saveGGplot}
\title{Save ggplot(s) as a PDF}
\usage{
saveGGplot(x, name, folder = NULL, width = 6, height = 4, multipage = FALSE)
}
\arguments{
\item{x}{A ggplot object or a list of ggplot objects}

\item{name}{A character string indicating the filename}

\item{folder}{A character string indicating the folder where the PDF should be saved}

\item{width}{Width of each panel; default = 10}

\item{height}{Height of each panel; default = 10}

\item{multipage}{A logical indicating if \code{x} is a list of plots to be saved as a multi-page PDF}
}
\description{
Save ggplot(s) as a PDF
}
\examples{
data <- data.frame(x = stats::runif(100, min = 0, max = 100),
y = stats::runif(100, min = 0, max = 100),
z = sample(c("A", "B", "C"), 100, replace = TRUE))
figure <- ggplot2::ggplot(data = data) +
ggplot2::geom_point(ggplot2::aes(x = .data$x, y = .data$y, color = .data$z)) +
ggplot2::scale_x_continuous("X variable") +
ggplot2::scale_y_continuous("Y variable") +
ggplot2::labs(title = "Title", subtitle = "Subtitle") +
ggthemes::scale_color_colorblind("Z variable")
\dontrun{saveGGplot(figure, name = "figure.pdf")}
figures <- list("figure 1" = figure,
"figure 2" = figure)
\dontrun{saveGGplot(figures, name = "figures.pdf", multipage = TRUE)}
}
