% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSarahTheme.R
\name{setSarahTheme}
\alias{setSarahTheme}
\title{Set the ggplot theme using a SarahB theme}
\usage{
setSarahTheme(theme, example = TRUE)
}
\arguments{
\item{theme}{A string indicating the name of the SarahB theme}

\item{example}{A logical indicating if an example ggplot using the theme should be shown}
}
\description{
Most SarahB package themes require Google fonts. If the theme requires a Google font, the Google font name will be the SarahB ggplot theme name (i.e., the 'Barlow' theme requires the 'Barlow' Google font). Google fonts are freely available from (https://fonts.google.com/).
}
\details{
Available themes
\itemize{
\item Barlow: Requires the Barlow Google font (https://fonts.google.com/specimen/Barlow)
}
}
\examples{
setSarahTheme("Barlow")
}
