% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slug.R
\name{slug}
\alias{slug}
\title{Converts string text to slug}
\usage{
slug(x, allow.numbers = TRUE)
}
\arguments{
\item{x}{Character}

\item{allow.numbers}{TRUE/FALSE}
}
\value{
Slugged text
}
\description{
Converts string text to slug
}
\examples{
x <- "slug this 123"
slug(x)
}
