% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{Tidy environment}
\usage{
tidy(x = NULL, base.env = NULL)
}
\arguments{
\item{x}{A vector indicating the names of objects to tidy (remove); \code{NULL} will remove all objects that aren't in the base environment (\code{base.env}), if defined}

\item{base.env}{The base environment}
}
\description{
Tidy environment
}
\examples{
tidy()
}
