% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper}
\alias{wrapper}
\title{Wrap character strings to format paragraphs}
\usage{
wrapper(x, ...)
}
\arguments{
\item{x}{Vector}

\item{...}{}
}
\value{
Character vector
}
\description{
Each character string in the input is first split into paragraphs (or lines containing whitespace only). The paragraphs are then formatted by breaking lines at word boundaries. The target columns for wrapping lines and the indentation of the first and all subsequent lines of a paragraph can be controlled independently.
}
