% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRange.R
\name{ggRange}
\alias{ggRange}
\title{Range for continuous axes}
\usage{
ggRange(x, accuracy = 1, log = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{accuracy}{The number number to round to (e.g., '5', '0.1').}

\item{log}{A logical indicating if the range should be for a log scale. Accuracy parameter is ignored if TRUE.}
}
\value{
A vector with length two. The first element is the floored minimum and the second element is the ceiling-ed maximum.
}
\description{
Range for continuous axes
}
\examples{
x <- runif(10)
ggRange(x, accuracy = 0.01)
ggRange(x, accuracy = 0.01, log = TRUE)
}
