% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupColor.R
\name{groupColor}
\alias{groupColor}
\title{Shades of colors by groups}
\usage{
groupColor(
  n,
  names = NULL,
  palettes = NULL,
  color.and.fill = FALSE,
  show = FALSE
)
}
\arguments{
\item{n}{Number of shades needed by group (i.e., the number of items in each group)}

\item{names}{Names of shades (i.e., names of items)}

\item{palettes}{ggsci material design color palettes}

\item{color.and.fill}{A logical indicator denoting whether to return shades for both \code{fill} and \code{color} parameters (color shades are produced by darkening fill shades)}

\item{show}{A logical indicator of whether to show output colors}
}
\value{
A vector or list (if \code{color.and.fill = TRUE}) of color shades, with names if supplied
}
\description{
This function returns shades of colors by group, wherein items in Group A will have shades of red, items in Group B will have shades of yellow, and so forth.
}
\examples{
n <- c(2, 1, 2)
names <- c("roses", "apples", "oranges", "plants", "zucchini")
groupColor(n = n, names = names, show = TRUE)
groupColor(n = n, names = names, color.and.fill = TRUE, show = TRUE)
}
