% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupColorInfo.R
\name{groupColorInfo}
\alias{groupColorInfo}
\title{Extract \code{groupColor} parameters from data frame}
\usage{
groupColorInfo(df, group, subgroup)
}
\arguments{
\item{df}{Data frame}

\item{group}{Grouping variable for color family (e.g., reds)}

\item{subgroup}{Subgroup variable for shades (e.g., burgundy, cherry, scarlett)}
}
\value{
List of two with the number of unique color shades needed by group and the subgroup labels for each shade
}
\description{
This is a helper function that allows for easy extraction of the \code{n} and \code{names} parameters for the \code{groupColor} function from a data frame.
}
\examples{
group = c("Red things", "Red things", "Yellow things", "Green things", "Green things")
subgroup = c("roses", "apples", "oranges", "plants", "zucchini")
data = c(5, 31, 19, 93, 29)
df <- data.frame(group, subgroup, data)
groupColorInfo(df = df, group = "group", subgroup = "subgroup")
}
