% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{Load packages}
\usage{
loadPackages(pkg)
}
\arguments{
\item{pkg}{A vector of packages to be loaded}
}
\value{
Named logical indicating if packages were successfully loaded
}
\description{
Load packages
}
\examples{
pkg = c("countrycode", "ggplot2")
loadPackages(pkg)
}
