% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRound.R
\name{makeRound}
\alias{makeRound}
\title{Formatted round numbers}
\usage{
makeRound(x, digits = 0)
}
\arguments{
\item{x}{Numeric}

\item{digits}{Number of digits to round to}
}
\value{
Character
}
\description{
Formatted round numbers
}
\examples{
x <- sample(seq(0, 10000), 10)
makeRound(x)
}
