% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifyScript.R
\name{notifyScript}
\alias{notifyScript}
\title{Print a notification of a script's conclusion in the console}
\usage{
notifyScript(prefix = "Done:", scriptName = NULL)
}
\arguments{
\item{prefix}{"Done:"}

\item{scriptName}{Searches for "scriptName" in global environment, otherwise must be supplied}
}
\value{
Printed notification text
}
\description{
Print a notification of a script's conclusion in the console
}
\examples{
scriptName <- "script.R"
notifyScript()
}
