% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb.classifications.R
\docType{data}
\name{wb.classifications}
\alias{wb.classifications}
\title{World Bank classifications for FY 2021-2022}
\format{
A data frame with 217 rows and 7 variables:
\describe{
\item{iso3}{ISO 3-character code}
\item{country}{Country name per World Bank formatting}
\item{region}{World Bank region}
\item{region.abb}{World Bank region abbreviation}
\item{year}{Year of data}
\item{income}{World Bank income group}
\item{lending}{World Bank lending category}
}
}
\source{
\url{https://datatopics.worldbank.org/world-development-indicators/the-world-by-income-and-region.html}
}
\usage{
wb.classifications
}
\description{
A dataset containing World Bank country names, codes, regions, income groups, and lending categories.
}
\keyword{datasets}
