% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper}
\alias{wrapper}
\title{Wrap long character strings}
\usage{
wrapper(x, width)
}
\arguments{
\item{x}{A character vector}

\item{width}{A positive integer giving the maximum width for wrapping lines}
}
\value{
A character vector
}
\description{
Strings are formatted by inserting line breaks at word boundaries, such that all lines are less than the user's maximum desired width.
}
\examples{
x <- c("This is a super super super super duper long string",
"This is also a super super super super duper long string")
wrapper(x, width = 20)
}
